requirements.txt

streamlit>=1.40.0
emotiefflib[torch]>=0.0.6
facenet-pytorch>=0.2.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0